Fix the X11 capability.
The openface2_ros container would always crash. It turns out that it was
because it failed to open the windows it usually creates. The simple
`DISPLAY` environment variable was not enough to allow for the container
to interact with X11. One had to either have the UID and GID of the user
in the container match those of the user currently logged in on the
host, or follow the instructions here : https://stackoverflow.com/a/25280523

I chose to follow the latter, as it does not require the docker image to
be custom-built for each UID and GID. It, however, requires that one
runs a command on the host computer before starting the container :

```shell
xauth nlist $DISPLAY | sed -e 's/^..../ffff/' | xauth -f /tmp/.docker.xauth nmerge -
```